+++ R/glex.R
@@ -85,9 +85,9 @@ glex.rpf <- function(object, x, max_interaction = NULL, features = NULL, ...) {
 #' library(xgboost)
 #' x <- as.matrix(mtcars[, -1])
 #' y <- mtcars$mpg
-#' xg <- xgboost(data = x[1:26, ], label = y[1:26],
-#'               params = list(max_depth = 4, eta = .1),
-#'               nrounds = 10, verbose = 0)
+#' xg <- xgboost(x[1:26, ], y[1:26],
+#'               max_depth = 4, learning_rate = .1,
+#'               nrounds = 10, verbosity = 0, nthreads = 1)
 #' glex(xg, x[27:32, ])
 #' glex(xg, mtcars[27:32, ])
 #'