+++ R/glex_vi.R
@@ -32,9 +32,9 @@
 #' library(xgboost)
 #' x <- as.matrix(mtcars[, -1])
 #' y <- mtcars$mpg
-#' xg <- xgboost(data = x[1:26, ], label = y[1:26],
-#'               params = list(max_depth = 4, eta = .1),
-#'               nrounds = 10, verbose = 0)
+#' xg <- xgboost(x[1:26, ], y[1:26],
+#'               max_depth = 4, learning_rate = .1,
+#'               nrounds = 10, verbosity = 0, nthreads = 1)
 #' glex_xgb <- glex(xg, x[27:32, ])
 #' vi_xgb <- glex_vi(glex_xgb)
 #'