+++ tests/testthat/test-glex-dataframe.R
@@ -3,7 +3,7 @@ library(ranger)
 
 test_that("Explaining XGBoost model with dataframe input works", {
   x <- as.matrix(mtcars[, -1])
-  xg <- xgboost(x, mtcars$mpg, nrounds = 10, verbose = 0)
+  xg <- xgboost(x, mtcars$mpg, nrounds = 10, verbosity = 0)
   glex_matrix <- glex(xg, x)
   glex_dataframe <- glex(xg, mtcars)
   expect_equal(glex_matrix$m, glex_dataframe$m)