+++ NEWS.md
@@ -1,4 +1,4 @@
-# glex (development version)
+# glex 0.6.0
 
 * Extended compatibility with `xgboost`, now requiring `xgboost (>= 3.0.0)` in `Suggests:`
   - Updated tests and examples for the new API